+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Call `malloc_trim(0)` after `System.gc()` on Linux to make sure memory gets released ([issue bytedeco/javacpp-presets#423](https://github.com/bytedeco/javacpp-presets/issues/423))
  * Make public the `Pointer.formatBytes()` and `Pointer.parseBytes()` static methods
  * Use `Integer.decode()` instead of `parseInt()` on integer literals to support hexadecimal and octal numbers
  * Add `Builder.encoding` option to let users specify I/O character set name ([issue bytedeco/javacpp-presets#195](https://github.com/bytedeco/javacpp-presets/issues/195))